+++ src/main/java/com/github/mjkuranda/spaceadventure2/SpaceAdventure2.java
@@ -2,7 +2,6 @@ package com.github.mjkuranda.spaceadventure2;
 
 import com.github.mjkuranda.spaceadventure2.states.GameState;
 import com.github.mjkuranda.spaceadventure2.states.IntroState;
-import org.lwjgl.opengl.Display;
 import org.newdawn.slick.*;
 import org.newdawn.slick.state.StateBasedGame;
 
@@ -20,8 +19,8 @@ public class SpaceAdventure2 extends StateBasedGame {
 
     @Override
     public void initStatesList(GameContainer container) throws SlickException {
-        addState(new IntroState());
         addState(new GameState());
+        addState(new IntroState());
     }
 
     public static void main(String[] args) {
@@ -32,7 +31,8 @@ public class SpaceAdventure2 extends StateBasedGame {
         try {
             appgc = new AppGameContainer(sg);
             appgc.setDisplayMode(screenSize.width, screenSize.height, false);
-            appgc.setShowFPS(false);
+            appgc.setShowFPS(true);
+            appgc.setTargetFrameRate(60);
 
             appgc.start();
         } catch (SlickException ex) {