+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/EntityType.java
@@ -2,6 +2,7 @@ package com.github.mjkuranda.spaceadventure2.entities;
 
 public enum EntityType {
     NONE,
+    SPACESHIP,
 
     /** Space entity list */
     ASTEROID,