+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/shootable/Shooting.java
@@ -1,22 +1,25 @@
 package com.github.mjkuranda.spaceadventure2.entities.shootable;
 
+import com.github.mjkuranda.spaceadventure2.entities.Entity;
 import com.github.mjkuranda.spaceadventure2.entities.missiles.Missile;
+import com.github.mjkuranda.spaceadventure2.entities.missiles.MissileType;
 
 import java.util.LinkedList;
 
 public class Shooting implements Shootable {
 
-    private LinkedList<Missile> subscriber;
+    private LinkedList<Entity> subscriber;
 
-    private Missile missile;
+    private MissileType missileType;
 
-    public Shooting(LinkedList<Missile> subscriber, Missile missile) {
+    public Shooting(LinkedList<Entity> subscriber, MissileType missileType) {
         this.subscriber = subscriber;
-        this.missile = missile;
+        this.missileType = missileType;
     }
 
     @Override
     public void shoot() {
-        subscriber.add(missile);
+        // FIXME
+        subscriber.add(null);
     }
 }