+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/GameState.java
@@ -1,10 +1,24 @@
 package com.github.mjkuranda.spaceadventure2.states;
 
+import com.github.mjkuranda.spaceadventure2.map.GameMap;
+import com.github.mjkuranda.spaceadventure2.renderers.ArcadeRenderer;
+import com.github.mjkuranda.spaceadventure2.renderers.Renderer;
 import org.newdawn.slick.*;
 import org.newdawn.slick.state.BasicGameState;
 import org.newdawn.slick.state.StateBasedGame;
 
 public class GameState extends BasicGameState {
+
+    /** Game map */
+    private GameMap map;
+
+    /** Game map renderer */
+    private Renderer renderer;
+
+    public GameState() {
+        map = new GameMap();
+        renderer = new ArcadeRenderer(map);
+    }
 
     @Override
     public int getID() {
@@ -18,11 +32,13 @@ public class GameState extends BasicGameState {
 
     @Override
     public void render(GameContainer container, StateBasedGame game, Graphics g) throws SlickException {
-
+        renderer.render(g);
     }
 
     @Override
     public void update(GameContainer container, StateBasedGame game, int delta) throws SlickException {
+        map.update(container, game);
+
         Input in = container.getInput();
 
         if (in.isKeyPressed(Input.KEY_F)) {